requirements.txt
# local packages if available
-e .

# set the project root as working directory for all jupyter notebooks 
# external requirements
--extra-index-url https://test.pypi.org/simple/projectroot
ProjectRoot

# standard libs
numpy<1.18
scipy==1.4.1
jupyterlab
matplotlib
pandas
pyyaml
ipywidgets
simpleitk
seaborn
tensorflow
sklearn
scikit-image
albumentations
nibabel
tqdm
